+++ packages/amplify-category-auth/provider-utils/awscloudformation/assets/string-maps.js
@@ -302,8 +302,8 @@ const coreAttributes = [
     name: 'Locale',
     value: 'locale',
   }, {
-    name: 'Middle Name',
-    value: 'middle_name',
+    name: 'Given Name',
+    value: 'given_name',
   },
   {
     name: 'Name',
@@ -432,7 +432,7 @@ const oAuthFlows = [
   },
   {
     name: 'Implicit grant',
-    value: 'token',
+    value: 'implicit',
   },
 ];