+++ packages/amplify-e2e-tests/src/init/amplifyPush.ts
@@ -1,7 +1,7 @@
 import * as nexpect from 'nexpect';
 import { getCLIPath, isCI } from '../utils';
 
-export default function amplifyPush(
+function amplifyPush(
   cwd: string,
   verbose: Boolean = isCI() ? false : true
 ) {
@@ -22,3 +22,25 @@ export default function amplifyPush(
       });
   });
 }
+
+function amplifyPushAuth(
+  cwd: string,
+  verbose: Boolean = isCI() ? false : true
+) {
+  return new Promise((resolve, reject) => {
+    nexpect
+      .spawn(getCLIPath(), ['push'], { cwd, stripColors: true, verbose })
+      .wait('Are you sure you want to continue?')
+      .sendline('y')
+      .wait(/.*/)
+      .run(function(err: Error) {
+        if (!err) {
+          resolve();
+        } else {
+          reject(err);
+        }
+      });
+  });
+}
+
+export { amplifyPush, amplifyPushAuth };