+++ packages/amplify-e2e-tests/src/utils/index.ts
@@ -3,6 +3,7 @@ import { mkdirSync } from 'fs';
 import * as rimraf from 'rimraf';
 import { config } from 'dotenv';
 export { default  as getProjectMeta } from './projectMeta';
+export { getUserPool, getUserPoolClients } from './sdk-calls'
 
 // run dotenv config to update env variable
 config();