+++ packages/amplify-graphql-docs-generator/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-docs-generator@1.3.1...amplify-graphql-docs-generator@1.4.1) (2019-04-09)
+
+**Note:** Version bump only for package amplify-graphql-docs-generator
+
+
+
+
+
 ## [1.3.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-docs-generator@1.0.8...amplify-graphql-docs-generator@1.3.1) (2019-04-03)
 
 **Note:** Version bump only for package amplify-graphql-docs-generator